# src/utils/swagger/swagger.json
{
	"basePath": "",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IAuthorDto": {
			"properties": {
				"id": {
					"type": "string",
					"x-nullable": true
				},
				"name": {
					"type": "string"
				}
			},
			"required": [
				"name"
			],
			"type": "object"
		},
		"IBookDto": {
			"properties": {
				"id": {
					"type": "string",
					"x-nullable": true
				},
				"title": {
					"type": "string"
				},
				"description": {
					"type": "string",
					"x-nullable": true
				},
				"dateOfPublication": {
					"type": "string"
				},
				"authorId": {
					"type": "string"
				},
				"author": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"title",
				"dateOfPublication",
				"authorId"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "ts-base-stack",
		"version": "1.0.0",
		"description": "A simple typescript/tsao base stack for client applications",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/authors": {
			"get": {
				"operationId": "GetAuthors",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IAuthorDto"
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAuthor",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/IAuthorDto"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IAuthorDto"
						}
					}
				]
			}
		},
		"/authors/{id}": {
			"get": {
				"operationId": "GetAuthor",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IAuthorDto"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/books": {
			"get": {
				"operationId": "GetBooks",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IBookDto"
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBook",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/IBookDto"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IBookDto"
						}
					}
				]
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetBook",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IBookDto"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}